dashboard/requirements.txt
streamlit>=1.37.0
pandas>=2.1.0
plotly>=5.18.0
statsmodels>=0.14.0
openai>=1.30.0
azure-data-tables>=12.5.0
databricks-sql-connector>=3.0.0
sqlglot>=23.0.0
python-dotenv>=1.0.0
